requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-auth==2.26.2
idna==3.6
kubernetes==29.0.0
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
urllib3==2.1.0
websocket-client==1.7.0
